refactor: extract Places initialization to custom Application class
- Created NavigationSampleApplication

- Moved Places API initialization and metadata extraction logic from SplashScreenActivity to NavigationSampleApplication

- Registered NavigationSampleApplication in AndroidManifest.xml
